+++ packages/expo-modules-core/android/src/androidTest/java/expo/modules/kotlin/jni/JavaScriptAnonymousObjectTest.kt
@@ -27,7 +27,7 @@ class JavaScriptAnonymousObjectTest {
     val f1 = evaluateScript("object.f1()").getString()
     val f2 = waitForAsyncFunction(it, "object.f2()").getString()
 
-    Truth.assertThat(anonymousObject.getPropertyNames()).isEqualTo(arrayOf("f2", "f1", "p1"))
+    Truth.assertThat(anonymousObject.getPropertyNames()).isEqualTo(arrayOf("p1", "f2", "f1"))
     Truth.assertThat(p1).isEqualTo(123)
     Truth.assertThat(f1).isEqualTo("abc")
     Truth.assertThat(f2).isEqualTo("def")